requirements.txt
torch
torchaudio
opencv-python
ffmpeg-python
sentence-transformers
openai-whisper
numpy
